packages/training_worker/pyproject.toml
[project]
name = "larynx-training-worker"
version = "0.1.0"
description = "LoRA fine-tuning worker for the Larynx platform (M7)."
requires-python = ">=3.11"
dependencies = [
    "larynx-shared",
    # We invoke the upstream VoxCPM training script as a subprocess (see
    # ORCHESTRATION-M7.md §6.1 and the third_party/ note in §0), so the
    # worker package itself doesn't depend on torch / nano-vllm / flash-
    # attn directly. The subprocess ships with its own deps via the
    # [gpu] extra on the root project.
    "structlog>=24.1",
    "soundfile>=0.12",
    "redis>=5.0",
    "pyyaml>=6.0",
]

[project.scripts]
larynx-training-worker = "larynx_training_worker.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/larynx_training_worker"]
